… .pre-commit-config.yaml
… superset-frontend/PR_DESCRIPTION.md
… superset-frontend/nx.json
… superset-frontend/package.json
… superset-frontend/packages/generator-superset/package.json
… superset-frontend/packages/generator-superset/project.json
… superset-frontend/packages/superset-core/package.json
… superset-frontend/packages/superset-core/project.json
… superset-frontend/packages/superset-ui-chart-controls/project.json
… superset-frontend/packages/superset-ui-core/package.json
… superset-frontend/packages/superset-ui-core/project.json
… superset-frontend/packages/superset-ui-demo/package.json
… superset-frontend/packages/superset-ui-demo/project.json
… superset-frontend/packages/superset-ui-switchboard/package.json
… superset-frontend/packages/superset-ui-switchboard/project.json
… superset-frontend/plugins/legacy-plugin-chart-calendar/package.json
… superset-frontend/plugins/legacy-plugin-chart-calendar/project.json
… superset-frontend/plugins/legacy-plugin-chart-calendar/src/vendor/cal-heatmap.js
… superset-frontend/plugins/legacy-plugin-chart-chord/package.json
… superset-frontend/plugins/legacy-plugin-chart-chord/project.json
… superset-frontend/plugins/legacy-plugin-chart-country-map/package.json
… superset-frontend/plugins/legacy-plugin-chart-country-map/project.json
… superset-frontend/plugins/legacy-plugin-chart-horizon/package.json
… superset-frontend/plugins/legacy-plugin-chart-horizon/project.json
… superset-frontend/plugins/legacy-plugin-chart-map-box/package.json
… superset-frontend/plugins/legacy-plugin-chart-map-box/project.json
… superset-frontend/plugins/legacy-plugin-chart-paired-t-test/package.json
… superset-frontend/plugins/legacy-plugin-chart-paired-t-test/project.json
… superset-frontend/plugins/legacy-plugin-chart-parallel-coordinates/package.json
… superset-frontend/plugins/legacy-plugin-chart-parallel-coordinates/project.json