@rusackas 2026-02-10
Closing in favor of a fresh approach. The codebase has changed significantly since this PR was created, and we're now targeting Turborepo instead of Nx for build orchestration. Starting a new POC branch with Bun + Turborepo.